requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-multipart==0.0.20
langgraph==0.6.6
PyYAML==6.0.2
pypdf==5.9.0
python-docx==1.2.0

pytest==8.4.1
pytest-cov==6.2.1
pytest-xdist==3.8.0
httpx==0.28.1
anyio==4.10.0

ruff==0.12.11
black==25.1.0
mypy==1.17.1
